refactor config definitions + compile sync action will run dbt deps if needed
