--- src/ConfigDefinition/SyncAction/DbtCompileDefinition.php
@@ -1,138 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DbtTransformation\ConfigDefinition\SyncAction;
-
-use DbtTransformation\ConfigDefinition\ConfigDefinition;
-use DbtTransformation\DwhProvider\DwhProviderFactory;
-use Keboola\Component\Config\BaseConfigDefinition;
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
-
-class DbtCompileDefinition extends BaseConfigDefinition
-{
-    protected function getParametersDefinition(): ArrayNodeDefinition
-    {
-        $parametersNode = parent::getParametersDefinition();
-        // @formatter:off
-        /** @noinspection NullPointerExceptionInspection */
-        $parametersNode
-            ->isRequired()
-            ->children()
-                ->arrayNode('git')
-                    ->ignoreExtraKeys()
-                    ->validate()
-                        ->always(function ($item) {
-                            if ((empty($item['username']) && !empty($item['#password']))
-                                || (!empty($item['username']) && empty($item['#password']))
-                            ) {
-                                throw new InvalidConfigurationException('Both username and password has to be set.');
-                            }
-                            return $item;
-                        })
-                    ->end()
-                    ->isRequired()
-                    ->children()
-                        ->scalarNode('repo')
-                            ->isRequired()
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('branch')
-                        ->end()
-                        ->scalarNode('username')
-                        ->end()
-                        ->scalarNode('#password')
-                        ->end()
-                    ->end()
-                ->end()
-            ->end();
-
-        $parametersNode
-            ->children()
-                ->arrayNode('dbt')
-                    ->children()
-                        ->arrayNode('modelNames')
-                            ->scalarPrototype()->end()
-                        ->end()
-                        ->scalarNode('threads')
-                            ->defaultValue(4)
-                        ->end()
-                    ->end()
-                ->end()
-            ->end();
-
-        /** @noinspection NullPointerExceptionInspection */
-        $parametersNode
-            ->children()
-                ->arrayNode('remoteDwh')
-                    ->validate()
-                        ->ifTrue(function ($v) {
-                            if (!is_array($v)) {
-                                return true;
-                            }
-                            $requiredSettings = ConfigDefinition::getRemoteDwhConnectionParams($v['type']);
-                            foreach ($requiredSettings as $setting) {
-                                if (!array_key_exists($setting, $v)) {
-                                    return true;
-                                }
-                            }
-                            return false;
-                        })
-                        ->thenInvalid('Missing required options for "%s"')
-                    ->end()
-                    ->children()
-                        ->enumNode('type')
-                            ->isRequired()
-                            ->values(DwhProviderFactory::REMOTE_DWH_ALLOWED_TYPES)
-                        ->end()
-                        ->scalarNode('host')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('server')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('user')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('#password')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('port')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('dbname')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('schema')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('warehouse')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('database')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('method')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('project')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('dataset')
-                            ->cannotBeEmpty()
-                        ->end()
-                            ->scalarNode('threads')
-                            ->defaultValue(4)
-                        ->end()
-                        ->scalarNode('#key_content')
-                            ->cannotBeEmpty()
-                        ->end()
-                    ->end()
-                ->end()
-            ->end();
-
-        // @formatter:on
-        return $parametersNode;
-    }
-}
