+++ src/Configuration/SyncAction/DbtCompileDefinition.php
@@ -2,12 +2,15 @@
 
 declare(strict_types=1);
 
-namespace DbtTransformation\ConfigDefinition\SyncAction;
+namespace DbtTransformation\Configuration\SyncAction;
 
+use DbtTransformation\Configuration\NodeDefinition\DbtNodeSyncAction;
+use DbtTransformation\Configuration\NodeDefinition\GitNode;
+use DbtTransformation\Configuration\NodeDefinition\RemoteDwhNode;
 use Keboola\Component\Config\BaseConfigDefinition;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 
-class DbtRunResultsDefinition extends BaseConfigDefinition
+class DbtCompileDefinition extends BaseConfigDefinition
 {
     protected function getParametersDefinition(): ArrayNodeDefinition
     {
@@ -17,11 +20,9 @@ class DbtRunResultsDefinition extends BaseConfigDefinition
         $parametersNode
             ->isRequired()
             ->children()
-                ->scalarNode('configId')
-                    ->isRequired()
-                ->end()
-                ->integerNode('branchId')
-                ->end()
+                ->append(new GitNode())
+                ->append(new DbtNodeSyncAction())
+                ->append(new RemoteDwhNode())
             ->end();
 
         // @formatter:on
