+++ tests/functionalSyncActions/run-sync-action-dbt-compile/source/data/config.json
@@ -13,7 +13,8 @@
   },
   "parameters": {
     "git": {
-      "repo": "https://github.com/keboola/dbt-test-project-public.git"
+      "repo": "https://github.com/keboola/dbt-test-project-public.git",
+      "branch": "branch-with-deps"
     },
     "dbt": {
       "modelNames": [
